feat: v0.7.2 preimage integration function (#474)

+++ internal/block/block_test.go
@@ -59,7 +59,7 @@ func Test_BlockEncodeDecode(t *testing.T) {
 
 	preimageExtrinsic := PreimageExtrinsic{
 		{
-			ServiceIndex: uint32(1),
+			ServiceIndex: 1,
 			Data:         []byte("preimage data"),
 		},
 	}
